requirements.txt
matplotlib
tensorflow_datasets
tensorflow==2.2.0
opencv_python
numpy
pandas
PyYAML
